@@ -18,35 +19,28 @@
     "pree2e": "webdriver-manager update --standalone false --gecko false",
     "e2e": "protractor"
   },
-  "private": true,
   "dependencies": {
-
-    "@angular/common": "2.2.3",
-    "@angular/compiler": "2.2.3",
-    "@angular/core": "2.2.3",
-    "@angular/forms": "2.2.3",
-    "@angular/http": "2.2.3",
-    "@angular/platform-browser": "2.2.3",
-    "@angular/platform-browser-dynamic": "2.2.3",
-    "@angular/platform-server": "2.2.3",
-    "@angular/router": "3.2.3",
-
+    "@angular/common": "2.4.4",
+    "@angular/compiler": "2.4.4",
+    "@angular/core": "2.4.4",
+    "@angular/forms": "2.4.4",
+    "@angular/http": "2.4.4",
+    "@angular/platform-browser": "2.4.4",
+    "@angular/platform-browser-dynamic": "2.4.4",
+    "@angular/platform-server": "2.4.4",
+    "@angular/router": "3.4.4",
     "@angularclass/conventions-loader": "^1.0.2",
     "@angularclass/hmr": "~1.2.2",
     "@angularclass/hmr-loader": "~3.0.2",
-
     "amcharts3": "github:amcharts/amcharts3",
     "ammap3": "github:amcharts/ammap3",
     "animate.css": "^3.5.1",
-    "assets-webpack-plugin": "^3.4.0",
     "bootstrap": "4.0.0-alpha.4",
-    "bootstrap-loader": "2.0.0-beta.12",
     "chart.js": "^1.1.1",
-    "chartist": "^0.9.7",
+    "chartist": "^0.10.1",
     "ckeditor": "4.6.0",
     "core-js": "^2.4.1",
     "easy-pie-chart": "^2.1.7",
-    "extract-text-webpack-plugin": "2.0.0-beta.3",
     "font-awesome": "^4.6.3",
     "font-awesome-sass-loader": "^1.0.2",
     "fullcalendar": "^2.7.2",
@@ -60,71 +54,78 @@
     "leaflet": "^0.7.7",
     "leaflet-map": "^0.2.1",
     "lodash": "^4.12.0",
-    "ng2-bootstrap": "1.1.16",
-    "ng2-ckeditor": "1.0.7",
-    "ng2-smart-table": "^0.3.2",
-    "ng2-tree": "^0.0.2-7",
-    "ng2-uploader": "1.1.0",
+    "ng2-bootstrap": "^1.3.0",
+    "ng2-ckeditor": "1.1.5",
+    "ng2-smart-table": "^0.5.1-0",
+    "ng2-tree": "^2.0.0-alpha.0",
+    "ngx-uploader": "^2.0.24",
     "normalize.css": "^4.1.1",
-    "postcss-loader": "^1.0.0",
-    "rxjs": "5.0.0-beta.12",
+    "reflect-metadata": "^0.1.9",
+    "rxjs": "~5.0.2",
     "tether": "^1.2.4",
-    "zone.js": "~0.7.1"
+    "zone.js": "~0.7.4"
   },
   "devDependencies": {
-    "angular2-router-loader": "^0.3.4",
-    "angular2-template-loader": "^0.6.0",
+    "@angular/compiler-cli": "2.4.4",
     "@types/electron": "^1.3.20",
+    "@types/fullcalendar": "^2.7.35",
     "@types/hammerjs": "^2.0.33",
-    "@types/jasmine": "^2.2.34",
+    "@types/jquery ": "^2.0.33",
+    "@types/jquery.slimscroll": "^1.3.30",
+    "@types/lodash": "0.0.28",
     "@types/node": "^6.0.38",
     "@types/source-map": "^0.1.27",
     "@types/uglify-js": "^2.0.27",
-    "@types/webpack": "2.0.0",
-    "@types/lodash": "0.0.28",
-    "@types/jquery ": "^2.0.33",
-    "@types/fullcalendar": "^2.7.35",
-    "@types/jquery.slimscroll": "^1.3.30",
-    "gh-pages": "^0.11.0",
-    "parse5": "^2.2.2",
-    "rimraf": "~2.5.4",
-    "script-ext-html-webpack-plugin": "^1.3.2",
-    "codelyzer": "~1.0.0-beta.4",
-    "tslint": "3.15.1",
-    "ts-helpers": "1.1.2",
-    "ts-node": "^1.7.0",
-    "typedoc": "^0.5.0",
-    "typescript": "2.0.3",
-    "awesome-typescript-loader": "^2.2.1",
-    "tslint-loader": "^2.1.3",
-    "url-loader": "^0.5.7",
-    "style-loader": "^0.13.1",
-    "raw-loader": "0.5.1",
-    "source-map-loader": "^0.1.5",
-    "string-replace-loader": "1.0.5",
-    "imports-loader": "^0.6.5",
-    "json-loader": "^0.5.4",
-    "css-loader": "^0.25.0",
+    "@types/webpack": "^2.0.0",
+    "add-asset-html-webpack-plugin": "^1.0.2",
+    "angular2-template-loader": "^0.6.0",
+    "assets-webpack-plugin": "^3.5.1",
+    "awesome-typescript-loader": "~3.0.0-beta.18",
+    "bootstrap-loader": "2.0.0-beta.18",
+    "codelyzer": "~2.0.0-beta.4",
+    "compression-webpack-plugin": "^0.3.1",
+    "copy-webpack-plugin": "^4.0.1",
+    "css-loader": "^0.26.1",
+    "electron": "^1.4.0",
     "exports-loader": "^0.6.3",
     "expose-loader": "^0.7.1",
+    "extract-text-webpack-plugin": "~2.0.0-beta.5",
     "file-loader": "^0.9.0",
-    "to-string-loader": "^1.1.4",
-    "sass-loader": "^4.0.2",
+    "find-root": "^1.0.0",
+    "gh-pages": "^0.11.0",
+    "html-webpack-plugin": "^2.26.0",
+    "imports-loader": "^0.7.0",
+    "json-loader": "^0.5.4",
+    "ng-router-loader": "^2.1.0",
+    "ngc-webpack": "^1.1.3",
+    "node-sass": "^4.3.0",
+    "npm-run-all": "^4.0.0",
+    "optimize-js-plugin": "0.0.4",
+    "parse5": "^3.0.1",
+    "postcss-loader": "^1.0.0",
+    "raw-loader": "0.5.1",
     "resolve-url-loader": "^1.4.3",
-    "node-sass": "^3.5.3",
-    "html-webpack-plugin": "^2.21.0",
-    "copy-webpack-plugin": "^3.0.1",
-    "v8-lazy-parse-webpack-plugin": "^0.3.0",
-    "webpack": "2.1.0-beta.27",
-    "webpack-dashboard": "^0.1.8",
-    "webpack-dev-middleware": "^1.6.1",
-    "webpack-dev-server": "2.1.0-beta.11",
-    "webpack-md5-hash": "^0.0.5",
-    "webpack-merge": "^0.17.0",
-    "compression-webpack-plugin": "^0.3.1",
+    "rimraf": "~2.5.4",
+    "sass-loader": "^5.0.0",
+    "script-ext-html-webpack-plugin": "^1.5.0",
+    "source-map-loader": "^0.1.6",
+    "string-replace-loader": "1.0.5",
+    "style-loader": "^0.13.1",
+    "to-string-loader": "^1.1.4",
+    "ts-helpers": "1.1.2",
+    "ts-node": "^2.0.0",
+    "tslint": "~4.4.0",
+    "typedoc": "^0.5.3",
+    "typescript": "~2.1.5",
+    "url-loader": "^0.5.7",
+    "webpack": "2.2.0",
+    "webpack-dashboard": "0.2.1",
+    "webpack-dev-middleware": "1.9.0",
+    "webpack-dev-server": "2.2.0",
+    "webpack-dll-bundles-plugin": "^1.0.0-beta.5",
+    "webpack-merge": "2.6.0",
     "es6-promise": "^3.1.2",
     "es6-shim": "^0.35.0",
-    "es7-reflect-metadata": "^1.6.0",
-    "electron": "^1.4.0"
+    "es7-reflect-metadata": "^1.6.0"
   }
 }
