@@ -1,9 +1,13 @@
 const path = require('path');
-
+const EVENT = process.env.npm_lifecycle_event || '';
 const ROOT = path.resolve(__dirname, '..');
 
 function hasProcessFlag(flag) {
     return process.argv.join('').indexOf(flag) > -1;
+}
+
+function hasNpmFlag(flag) {
+    return EVENT.includes(flag);
 }
 
 function isWebpackDevServer() {
@@ -15,5 +19,6 @@ function root(...args) {
 }
 
 exports.hasProcessFlag = hasProcessFlag;
+exports.hasNpmFlag = hasNpmFlag;
 exports.isWebpackDevServer = isWebpackDevServer;
 exports.root = root;
