@@ -5,13 +5,15 @@ const AssetsPlugin = require('assets-webpack-plugin');
 const NormalModuleReplacementPlugin = require('webpack/lib/NormalModuleReplacementPlugin');
 const ContextReplacementPlugin = require('webpack/lib/ContextReplacementPlugin');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
-const ForkCheckerPlugin = require('awesome-typescript-loader').ForkCheckerPlugin;
+const CheckerPlugin = require('awesome-typescript-loader').CheckerPlugin;
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const LoaderOptionsPlugin = require('webpack/lib/LoaderOptionsPlugin');
 const CommonsChunkPlugin = require('webpack/lib/optimize/CommonsChunkPlugin');
 const ScriptExtHtmlWebpackPlugin = require('script-ext-html-webpack-plugin');
+const ngcWebpack = require('ngc-webpack');
 
+const AOT = helpers.hasNpmFlag('aot');
 const METADATA = {
     title: 'Joker',
     description: 'Fight against landlords',
@@ -25,7 +27,7 @@ module.exports = function (options) {
         entry: {
             polyfills: './src/polyfills.browser.ts',
             vendor: './src/vendor.browser.ts',
-            main: './src/main.browser.ts'
+            main: AOT ? './src/main.browser.aot.ts' : './src/main.browser.ts'
         },
         resolve: {
             extensions: ['.ts', '.js', '.css', '.scss', '.json'],
@@ -35,21 +37,32 @@ module.exports = function (options) {
             rules: [
                 {
                     test: /\.ts$/,
-                    loader: 'string-replace-loader',
-                    query: {
-                        search: /(System|SystemJS)(.*[\n\r]\s*\.|\.)import\((.+)\)/g,
-                        replace: '$1.import($3).then(mod => (mod.__esModule && mod.default) ? mod.default : mod)'
-                    },
-                    include: [helpers.root('src')],
-                    enforce: 'pre'
-                },
-                {
-                    test: /\.ts$/,
                     use: [
-                        `@angularclass/hmr-loader?pretty=${!isProd}&prod=${isProd}`,
-                        'awesome-typescript-loader',
-                        'angular2-template-loader',
-                        'angular2-router-loader'
+                        {
+                            loader: '@angularclass/hmr-loader',
+                            options: {
+                                pretty: !isProd,
+                                prod: isProd
+                            }
+                        },
+                        {
+                            loader: 'ng-router-loader',
+                            options: {
+                                loader: 'async-import',
+                                genDir: 'compiled',
+                                aot: AOT
+                            }
+                        },
+                        {
+                            loader: 'awesome-typescript-loader',
+                            options: {
+                                configFileName: 'tsconfig.webpack.json'
+                            }
+                        },
+                        {
+                            loader: 'angular2-template-loader'
+                        }
+
                     ],
                     exclude: [/\.(spec|e2e)\.ts$/]
                 },
@@ -95,9 +108,7 @@ module.exports = function (options) {
                 }
             ]
         },
-        resolveLoader: {
-            moduleExtensions: ['-loader']
-        },
+
         plugins: [
             new ExtractTextPlugin({
                 filename: 'initial.css',
@@ -108,22 +119,28 @@ module.exports = function (options) {
                 filename: 'webpack-assets.json',
                 prettyPrint: true
             }),
-            new ForkCheckerPlugin(),
+            new CheckerPlugin(),
+            new CommonsChunkPlugin({
+                name: 'polyfills',
+                chunks: ['polyfills']
+            }),
+            new CommonsChunkPlugin({
+                name: 'vendor',
+                chunks: ['main'],
+                minChunks: module => /node_modules/.test(module.resource)
+            }),
             new CommonsChunkPlugin({
                 name: ['polyfills', 'vendor'].reverse()
             }),
+
             new ContextReplacementPlugin(
                 /angular(\\|\/)core(\\|\/)src(\\|\/)linker/,
                 helpers.root('src')
             ),
             new CopyWebpackPlugin([
-                {
-                    from: 'src/assets',
-                    to: 'assets'
-                },
-                {
-                    from: 'src/meta'
-                }
+                {from: 'src/assets', to: 'assets'},
+                {from: 'node_modules/ckeditor', to: 'ckeditor'},
+                {from: 'src/meta'}
             ]),
             new HtmlWebpackPlugin({
                 template: 'src/index.html',
@@ -173,7 +190,14 @@ module.exports = function (options) {
             new NormalModuleReplacementPlugin(
                 /facade(\\|\/)math/,
                 helpers.root('node_modules/@angular/core/src/facade/math.js')
-            )
+            ),
+
+            new ngcWebpack.NgcWebpackPlugin({
+                disabled: !AOT,
+                tsConfig: helpers.root('tsconfig.webpack.json'),
+                resourceOverride: helpers.root('config/resource-override.js')
+            })
+
         ],
         node: {
             global: true,
