@@ -1,9 +1,11 @@
 const helpers = require('./helpers');
 const webpackMerge = require('webpack-merge');
+const webpackMergeDll = webpackMerge.strategy({plugins: 'replace'});
 const commonConfig = require('./webpack.common.js');
-
+const AddAssetHtmlPlugin = require('add-asset-html-webpack-plugin');
 const DefinePlugin = require('webpack/lib/DefinePlugin');
 const LoaderOptionsPlugin = require('webpack/lib/LoaderOptionsPlugin');
+const DllBundlesPlugin = require('webpack-dll-bundles-plugin').DllBundlesPlugin;
 
 const ENV = process.env.ENV = process.env.NODE_ENV = 'development';
 const HOST = process.env.HOST || 'localhost';
@@ -55,7 +57,44 @@ module.exports = function (options) {
                         resourcePath: 'src'
                     }
                 }
-            })
+            }),
+            new DllBundlesPlugin({
+                bundles: {
+                    polyfills: [
+                        'core-js',
+                        {
+                            name: 'zone.js',
+                            path: 'zone.js/dist/zone.js'
+                        },
+                        {
+                            name: 'zone.js',
+                            path: 'zone.js/dist/long-stack-trace-zone.js'
+                        },
+                        'ts-helpers'
+                    ],
+                    vendor: [
+                        '@angular/platform-browser',
+                        '@angular/platform-browser-dynamic',
+                        '@angular/core',
+                        '@angular/common',
+                        '@angular/forms',
+                        '@angular/http',
+                        '@angular/router',
+                        '@angularclass/hmr',
+                        'rxjs'
+                    ]
+                },
+                dllDir: helpers.root('dll'),
+                webpackConfig: webpackMergeDll(commonConfig({env: ENV}), {
+                    devtool: 'cheap-module-source-map',
+                    plugins: []
+                })
+            }),
+
+            new AddAssetHtmlPlugin([
+                {filepath: helpers.root(`dll/${DllBundlesPlugin.resolveFile('polyfills')}`)},
+                {filepath: helpers.root(`dll/${DllBundlesPlugin.resolveFile('vendor')}`)}
+            ])
         ],
         devServer: {
             port: METADATA.port,
