@@ -1,8 +1,6 @@
-import {Routes, RouterModule} from '@angular/router';
-
+import { ModuleWithProviders } from '@angular/core';
 const ROUTES: Routes = [
     {path: '', redirectTo: 'work', pathMatch: 'full'},
     {path: '**', redirectTo: 'work/index'}
 ];
-
-export const routes = RouterModule.forRoot(ROUTES, {useHash: false});
+export const routes: ModuleWithProviders = RouterModule.forRoot(ROUTES, { useHash: true });
