@@ -2,13 +2,12 @@ import {Component, ViewEncapsulation} from '@angular/core';
 
 import {PieChartService} from './pieChart.service';
 
-import './pieChart.loader.ts';
+import 'easy-pie-chart/dist/jquery.easypiechart.js';
+import 'style-loader!./pieChart.scss';
 
 @Component({
   selector: 'pie-chart',
-  encapsulation: ViewEncapsulation.None,
-  styles: [require('./pieChart.scss')],
-  template: require('./pieChart.html')
+  templateUrl: './pieChart.html'
 })
 // TODO: move easypiechart to component
 export class PieChart {
