@@ -1,7 +1,11 @@
-import {Component, ViewEncapsulation} from '@angular/core';
+import { Component } from '@angular/core';
+import { Routes } from '@angular/router';
+
+import { BaMenuService } from '../theme';
+import { MENU } from '../app.menu';
+
 @Component({
   selector: 'pages',
-  encapsulation: ViewEncapsulation.None,
   styles: [],
   template: `
     <ba-sidebar></ba-sidebar>
@@ -29,9 +33,9 @@ import {Component, ViewEncapsulation} from '@angular/core';
 })
 export class Pages {
 
-  constructor() {
-  }
+  constructor(private _menuService: BaMenuService,) {  }
 
   ngOnInit() {
+    this._menuService.updateMenuByRoutes(<Routes>MENU);
   }
 }
