@@ -4,8 +4,8 @@ import {GlobalState} from '../../../global.state';
 
 @Component({
   selector: 'ba-content-top',
-  styles: [require('./baContentTop.scss')],
-  template: require('./baContentTop.html'),
+  styleUrls: ['./baContentTop.scss'],
+  templateUrl: './baContentTop.html',
 })
 export class BaContentTop {
 
