@@ -5,8 +5,8 @@ import {BaMsgCenterService} from './baMsgCenter.service';
 @Component({
   selector: 'ba-msg-center',
   providers: [BaMsgCenterService],
-  styles: [require('./baMsgCenter.scss')],
-  template: require('./baMsgCenter.html')
+  styleUrls: ['./baMsgCenter.scss'],
+  templateUrl: './baMsgCenter.html'
 })
 export class BaMsgCenter {
 
