@@ -2,7 +2,7 @@ import {NgModule, ModuleWithProviders}      from '@angular/core';
 import {CommonModule}  from '@angular/common';
 import {ReactiveFormsModule, FormsModule} from '@angular/forms';
 import {RouterModule} from '@angular/router';
-
+import { NgUploaderModule } from 'ngx-uploader';
 import {
     BaThemeConfig
 } from './theme.config';
@@ -107,6 +107,7 @@ const NGA_VALIDATORS = [
         RouterModule,
         FormsModule,
         ReactiveFormsModule,
+        NgUploaderModule
     ],
     exports: [
         ...NGA_PIPES,
