@@ -1,12 +1,13 @@
-import {Injectable} from '@angular/core';
-import {BaThemeConfigProvider} from './theme.configProvider';
-import {colorHelper} from './theme.constants';
+import { Injectable } from '@angular/core';
+
+import { BaThemeConfigProvider } from './theme.configProvider';
+import { colorHelper } from './theme.constants';
+
 
 @Injectable()
 export class BaThemeConfig {
 
-  constructor(private _baConfig:BaThemeConfigProvider) {
-    this._config();
+  constructor(private _baConfig: BaThemeConfigProvider) {
   }
 
   private _config() {
